NHANES data, 1999-2020

The NHANES program was initiated in the early 1960s and beginning in 1999 has been conducted continuously, in two-year cycles. The protocols for each cycle were approved by the NCHS Institutional Review Board. Written informed consent was obtained from each participant.

Usage

nhanes_data

Format

An object of class data.table (inherits from data.frame) with 63041 rows and 98 columns.

Source

https://wwwn.cdc.gov/nchs/nhanes/continuousnhanes/default.aspx

Details

NHANES data include in-home interviews and study examinations conducted at mobile examination centers. The interview included questions about demographics, health behaviors, prior diagnoses, medication use, and medical history. During the interview, the labels of medications that participants reported taking in the preceding 30 days were recorded. During the study examination, height and weight were measured, blood samples were used to measure cholesterol, glycated hemoglobin and serum creatinine, and a spot urine sample was used to measure albumin and creatinine and to conduct a pregnancy test.

Survey variables

In each NHANES cycle, potential participants were identified using a multi-stage sampling process. The variables below are based on this process.

svy_id

Label: participant identifier

Description: NHANES participant unique identifier.

svy_psu

Label: primary sampling unit

Description: Population sampling unit. This variable is used to account for the non-random selection of study participants for NHANES

svy_strata

Label: strata

Description: Population stratification. This variable is used to account for the non-random selection of study participants for NHANES

svy_weight_mec

Label: Mobile examination center weights

Description: Weight applied to produce statistical estimates for the non-institutionalized US population. This weight is used for calculating means and proportions.

svy_subpop_htn

Label: Subpopulation for hypertension

Description: This indicates that the person has data needed to be included in the analysis of blood pressure or hypertension data (i.e., they had at least one systolic and diastolic blood pressure measurement and they had information on self-reported antihypertensive medication use).

svy_year

Label: NHANES cycle

Description: NHANES survey cycle: 1999-2000, 2001-2002, 2003-2004, 2005-2006, 2007-2008, 2009-2010, 2011-2012, 2013-2014, 2015-2016, 2017-2020

Demographics variables

The NHANES in-home interview included questions about participant demographics, including age, sex, race/ethnicity, and pregnancy status

demo_age_cat

Label: Age category, years

Description: Age grouping: 18-44, 45-64, 65-74, >= 75 years

demo_race

Label: Race

Description: Self-reported race/ethnicity. From 1999-2000 through 2009-2010 this was available as non-Hispanic White, non-Hispanic Black, Hispanic and other. From 2011-2012 through 2017-2020 this was available as non-Hispanic White, non-Hispanic Black, non-Hispanic Asian, Hispanic and other.

demo_age_years

Label: Age, years

Description: Participant age in years. Participants > 80 years of age are given an age of 80 years.

demo_pregnant

Label: Pregnant

Description: Pregnancy status. This is defined by either self-report of being pregnant or a positive pregnancy test conducted during the study visit.

demo_gender

Label: Gender

Description: Self-reported gender

Blood pressure variables

Systolic and diastolic BP (SBP and DBP, respectively) were measured three times by trained and certified physicians. BP was measured using a mercury sphygmomanometer from 1999-2000 through 2015-2016 and using an oscillometric device in 2017-2020. The oscillometric SBP and DBP values were calibrated to the mercury device. Mean SBP and DBP levels were computed over all available

bp_sys_mean

Label: Systolic blood pressure, mm Hg

Description: Mean systolic blood pressure in mm Hg. This is based on the average of up to 3 readings. Overall, >95% of participants with at least one systolic blood pressure reading had three readings. From 1999-2000 through 2015-2016, systolic blood pressure was measured using a mercury sphygmomanometer. In 2017-2020, systolic blood pressure was measured using an oscillometric device. The systolic blood pressure in 2017-2020 was calibrated to the mercury device by adding 1.5 mm Hg to the mean measured value.

bp_dia_mean

Label: Diastolic blood pressure, mm Hg

Description: Mean diastolic blood pressure in mm Hg. This is based on the average of up to 3 readings. Overall, >95% of participants with at least one diastolic blood pressure reading had three readings. From 1999-2000 through 2015-2016, diastolic blood pressure was measured using a mercury sphygmomanometer. In 2017-2020, diastolic blood pressure was measured using an oscillometric device. The diastolic blood pressure in 2017-2020 was calibrated to the mercury device by subtracting 1.0 mm Hg to the mean measured value.

bp_cat_meds_excluded

Label: Blood pressure category

Description: Systolic/diastolic blood pressure <120/80, 120-129/<80, 130-139/80-89, 140-159/90-99, >= 160/100. Participants were placed in the category associated with higher blood pressure (e.g., someone with systolic blood pressure of 150 mm Hg and diastolic blood pressure of 76 mm Hg was placed in the >= 140/90 mm Hg category)

bp_cat_meds_included

Label: Blood pressure category (including antihypertensive medication use)

Description: Systolic/diastolic blood pressure <120/80, 120-129/<80, 130-139/80-89, 140-159/90-99, >= 160/100. Participants taking antihypertensive medication were placed in a separate category. Participants were placed in the category associated with higher blood pressure (e.g., someone with systolic blood pressure of 150 mm Hg and diastolic blood pressure of 76 mm Hg was placed in the >= 140/90 mm Hg category)

bp_control_jnc7

Label: Blood pressure control (JNC7 guideline)

Description: Systolic and diastolic blood pressure controlled to the levels recommended in the JNC7 guideline, systolic blood pressure < 140 mm Hg and diastolic blood pressure < 90 mm Hg.

bp_control_accaha

Label: Blood pressure control (2017 ACC/AHA BP guideline)

Description: Systolic and diastolic blood pressure controlled to the levels recommended in the 2017 ACC/AHA BP guideline, systolic blood pressure < 130 mm Hg and diastolic blood pressure < 80 mm Hg except for those >= 65 years of age without diabetes, chronic kidney disease, history of cardiovascular disease or 10-year predicted ASCVD risk >= 10%. For this group, blood pressure control was defined as systolic blood pressure < 130 mm Hg

bp_uncontrolled_jnc7

Label: Uncontrolled BP (JNC7 guideline)

Description: Systolic blood pressure >= 140 mm Hg or diastolic blood pressure >= 90 mm Hg

bp_uncontrolled_accaha

Label: Uncontrolled BP (2017 ACC/AHA BP guideline)

Description: Systolic blood pressure >= 130 mm Hg or diastolic blood pressure >= 80 mm Hg except for individuals without diabetes, chronic kidney disease, history of cardiovascular disease or 10-year predicted ASCVD risk >= 10%. For this group, uncontrolled blood pressure was defined as systolic blood pressure >= 130 mm Hg

Hypertension variables

These variables are derived according to the 2017 American College of Cardiology/American Heart Association (ACC/AHA) BP guideline and the Seventh Joint National Committee (JNC7) BP guideline.

htn_jnc7

Label: Hypertension (JNC7 guideline)

Description: Hypertension defined by the JNC7 guideline, systolic blood pressure >= 140 mm Hg, diastolic blood pressure >= 90 mm Hg or self-reported antihypertensive medication use.

htn_accaha

Label: Hypertension (2017 ACC/AHA BP guideline)

Description: Hypertension defined by the 2017 ACC/AHA blood pressure guideline, systolic blood pressure >= 130 mm Hg, diastolic blood pressure >= 80 mm Hg or self-reported antihypertensive medication use.

htn_aware

Label: Awareness of hypertension

Description: Self-report of a prior diagnosis of antihypertensive medication.

htn_resistant_jnc7

Label: Resistant hypertension (JNC7 guideline definition)

Description: Taking 4 or more classes of antihypertensive medication, systolic blood pressure >= 140 mm Hg or diastolic blood pressure >= 90 mm Hg with the use of 3 classes of antihypertensive medication, or systolic blood pressure >= 130 mm Hg or diastolic blood pressure >= 80 mm Hg with the use of 3 classes of antihypertensive medication for those with diabetes or chronic kidney disease.

htn_resistant_accaha

Label: Resistant hypertension (2017 ACC/AHA BP guideline)

Description: Taking 4 or more classes of antihypertensive medication, systolic blood pressure >= 140 mm Hg or diastolic blood pressure >= 90 mm Hg with the use of 3 classes of antihypertensive medication, or systolic blood pressure >= 130 mm Hg or diastolic blood pressure >= 80 mm Hg with the use of 3 classes of antihypertensive medication for those <65 years of age and those >= 65 years of age with diabetes, chronic kidney disease or high cardiovascular risk defined by a history of cardiovascular disease or 10-year predicted risk >= 10% using the pooled cohort risk equations.

htn_resistant_accaha_thz

Label: Resistant hypertension (JNC7 guideline, requires thiazide diuretic)

Description: Taking 4 or more classes of antihypertensive medication, systolic blood pressure >= 140 mm Hg or diastolic blood pressure >= 90 mm Hg with the use of 3 classes of antihypertensive medication, or systolic blood pressure >= 130 mm Hg or diastolic blood pressure >= 80 mm Hg with the use of 3 classes of antihypertensive medication for those < 65 years of age and those >= 65 years of age with diabetes, chronic kidney disease or high cardiovascular risk defined by a history of cardiovascular disease or 10-year predicted risk >= 10% using the pooled cohort risk equations. To meet this definition of resistant hypertension, the participant had to be taking a thiazide diuretic.

htn_resistant_jnc7_thz

Label: Resistant hypertension (2017 ACC/AHA BP guideline, requires thiazide diuretic)

Description: Taking 4 or more classes of antihypertensive medication, systolic blood pressure >= 140 mm Hg or diastolic blood pressure >= 90 mm Hg with the use of 3 classes of antihypertensive medication, or systolic blood pressure >= 130 mm Hg or diastolic blood pressure >= 80 mm Hg with the use of 3 classes of antihypertensive medication for those with diabetes or chronic kidney disease. To meet this definition of resistant hypertension, the participant had to be taking a thiazide diuretic.

Antihypertensive medication variables

During the interview, the labels of medications that participants reported taking in the preceding 30 days were recorded.

bp_med_use

Label: Self-reported antihypertensive medication use

Description: Self-reported use of antihypertensive medication

bp_med_recommended_jnc7

Label: Antihypertensive medications recommended (JNC7 guideline)

Description: Systolic blood pressure >= 140 mm Hg or diastolic blood pressure >= 90 mm Hg; Systolic blood pressure >= 130 mm Hg or diastolic blood pressure >= 80 mm Hg for those with chronic kidney disease or diabetes. Those taking antihypertensive medications were considered to be recommended treatment by this definition.

bp_med_recommended_accaha

Label: Antihypertensive medications recommended (2017 ACC/AHA BP guideline)

Description: Systolic blood pressure >= 140 mm Hg or diastolic blood pressure >= 90 mm Hg; Systolic blood pressure >= 130 mm Hg or diastolic blood pressure >= 80 mm Hg for those with chronic kidney disease, diabetes, 10-year predicted atherosclerotic cardiovascular disease risk by the pooled cohort risk equations or age >= 65 years. Those taking antihypertensive medications were considered to be recommended treatment by this definition.

bp_med_n_class

Label: Number of antihypertensive medication classes

Description: Number of antihypertensive medication classes being taken based on the pill bottle review

bp_med_ace

Label: ACE inhibitors

Description: Taking an angiotensin converting enzyme inhibitor, defined using the pill bottle review. Drugs in this class included bnazepril, captopril, enalapril, fosinopril, lisonopril, moexipril, perindopril, quinapril, ramipril, trandolapril

bp_med_aldo

Label: Aldosterone antagonists

Description: Taking an aldosterone antagonist, defined using the pill bottle review. Drugs in this class included eplerenone, spironolactone.

bp_med_alpha

Label: Alpha-1 blockers

Description: Taking an alpha blocker, defined using the pill bottle review. Drugs in this class included doxazosin, prazosin, terazosin.

bp_med_angioten

Label: Angiotensin receptor blockers

Description: Taking an angiotensin receptor blocker, defined using the pill bottle review. Drugs in this class included candesartan, eprosartan, irbesartan, losartan, olmesartan, telmisartan, valsartan, azilsartan.

bp_med_beta

Label: Beta blockers

Description: Taking a beta blocker. Drugs in this class included acebutolol, atenolol, betaxolol, bisoprolol, carvedilol, labetalol, metoprolol, nadolol, nebivolol, pindolol, propranolol.

bp_med_central

Label: Central alpha1 agonist and other centrally acting agents

Description: Taking a centrally acting agents, defined using the pill bottle review. Drugs in this class included clonidine, guanabenz, guanfaacine, methyldopa, reserpine.

bp_med_ccb

Label: Calcium channel blockers

Description: Taking a calcium channel blocker, defined using the pill bottle review. Drugs in this class included amlodipine, diltiazem, felodipine, isradipine, nicardipine, nifedipine, nisoldipine, verapamil.

bp_med_diur_Ksparing

Label: Potassium sparing diuretics

Description: Taking a potassium-sparing diuretic, defined using the pill bottle review. Drugs in this class included amiloride, triamterene.

bp_med_diur_loop

Label: Loop diuretics

Description: Taking a loop diuretic, defined using the pill bottle review. Drugs in this class included bumetanide, furosemide, torsemide, ethacrynic acid.

bp_med_diur_thz

Label: Thiazide or thiazide-type diuretics

Description: Taking a thiazide diuretic, defined using the pill bottle review. Drugs in this class included bendroflumethiazide, chlorthalidone, chlorothiazide, hydrochlorothiazide, indapamide, metolazone, polythiazide.

bp_med_renin_inhibitors

Label: Direct renin inhibitors

Description: Taking a renin inhibitor, defined using the pill bottle review. Drugs in this class included aliskiren.

bp_med_vasod

Label: Direct vasodilators

Description: Taking a vasodilator, defined using the pill bottle review. Drugs in this class included hydralazine, minoxidil.

Comorbidities variables

These variables describe comorbid conditions.

cc_n_highrisk

Label: Number of high risk conditions

Description: Self-reported history of coronary heart disease, myocardial infarction, stroke or heart failure or 10-year predicted risk >= 10% estimated by the pooled cohort risk equations

cc_smoke

Label: Smoking status

Description: Self-reported current cigarette smoking

cc_bmi

Label: Body mass index, kg/m2

Description: Body mass index in kg/m2, estimated using the height and weight measured during the study examination.

cc_diabetes

Label: Prevalent diabetes

Description: HbA1c >= 6.5% or self-report of a prior diagnosis of diabetes with use of insulin or oral glucose-lowering medications.

cc_ckd

Label: Prevalent chronic kidney disease

Description: Estimated glomerular filtration rate < 60 ml/min/1.73 m2 or albumin-to-creatinine > 30 mg/g. Estimated glomerular filtration rate was calculated using the 2021 serum creatinine-based equation.

cc_cvd_mi

Label: History of myocardial infarction

Description: Self-reported history of myocardial infarction

cc_cvd_chd

Label: History of coronary heart disease

Description: Self-reported history of myocardial infarction or coronary heart disease

cc_cvd_stroke

Label: History of stroke

Description: Self-reported history of stroke

cc_cvd_ascvd

Label: History of ASCVD

Description: Self-reported history of coronary heart disease, myocardial infarction or stroke

cc_cvd_hf

Label: History of heart failure

Description: Self-reported history of heart failure

cc_cvd_any

Label: History of CVD

Description: Self-reported history of coronary heart disease, myocardial infarction, stroke or heart failure