NHANES data, 1999-2020

The NHANES program was initiated in the early 1960s and beginning in 1999 has been conducted continuously, in two-year cycles. The protocols for each cycle were approved by the NCHS Institutional Review Board. Written informed consent was obtained from each participant.

Usage

nhanes_data

Format

An object of class data.table (inherits from data.frame) with 59799 rows and 101 columns.

Source

https://wwwn.cdc.gov/nchs/nhanes/continuousnhanes/default.aspx

Details

NHANES data include in-home interviews and study examinations conducted at mobile examination centers. The interview included questions about demographics, health behaviors, prior diagnoses, medication use, and medical history. During the interview, the labels of medications that participants reported taking in the preceding 30 days were recorded. During the study examination, height and weight were measured, blood samples were used to measure cholesterol, glycated hemoglobin and serum creatinine, and a spot urine sample was used to measure albumin and creatinine and to conduct a pregnancy test.

Currently, two sub-populations are included in nhanes_data:

blood pressure and hypertension (n = 56,017)

lipids and cholesterol (n = 25,151)

To access a sub-population, simply filter nhanes_data to contain rows where the corresponding sub-population indicator is equal to 1, e.g.,

# for blood pressure and hypertension sub-population
nhanes_data[svy_subpop_htn == 1]

# for lipids and cholesterol sub-population
nhanes_data[svy_subpop_chol == 1]

Survey variables

In each NHANES cycle, potential participants were identified using a multi-stage sampling process. The variables below are based on this process.

svy_id

Label: Participant identifier

Description: NHANES participant unique identifier.

svy_psu

Label: Primary sampling unit

Description: Population sampling unit. This variable is used to account for the non-random selection of study participants for NHANES

svy_strata

Label: Strata

Description: Population stratification. This variable is used to account for the non-random selection of study participants for NHANES

svy_weight_mec

Label: Mobile examination center weights

Description: Weight applied to produce statistical estimates for the non-institutionalized US population. This weight is used for calculating means and proportions.

svy_subpop_htn

Label: Subpopulation for hypertension

Description: This indicates that the person has data needed to be included in the analysis of blood pressure or hypertension data (i.e., they had at least one systolic and diastolic blood pressure measurement and they had information on self-reported antihypertensive medication use).

svy_year

Label: NHANES cycle

Description: NHANES survey cycle: 1999-2000, 2001-2002, 2003-2004, 2005-2006, 2007-2008, 2009-2010, 2011-2012, 2013-2014, 2015-2016, 2017-2020. Each cycle is 2 years in length, apart from 2017-2020, which covers 2017 through March of 2020 (approximately 3.2 years)

Demographics variables

The NHANES in-home interview included questions about participant demographics, including age, sex, race/ethnicity, and pregnancy status

demo_age_cat

Label: Age category, years

Description: Age grouping: 18-44, 45-64, 65-74, >= 75 years

Number of observed values: 56,017 (100%)

Counts of observed values

18 to 44: 26,096 (46.6%)

45 to 64: 17,155 (30.6%)

65 to 74: 7,013 (12.5%)

75+: 5,753 (10.3%)

Number of missing values: 0 (0.00%)

demo_race

Label: Race/ethnicity

Description: Self-reported race/ethnicity. From 1999-2000 through 2009-2010 this was available as non-Hispanic White, non-Hispanic Black, Hispanic and other. From 2011-2012 through 2017-2020 this was available as non-Hispanic White, non-Hispanic Black, non-Hispanic Asian, Hispanic and other.

Number of observed values: 56,017 (100%)

Counts of observed values

Non-Hispanic White: 23,980 (42.8%)

Non-Hispanic Black: 12,355 (22.1%)

Non-Hispanic Asian: 2,946 (5.26%)

Hispanic: 14,520 (25.9%)

Other: 2,216 (3.96%)

Number of missing values: 0 (0.00%)

demo_age_years

Label: Age, years

Description: Participant age in years. Participants > 80 years of age are given an age of 80 years. This top-coding was changed from 85 years to 80 with the 2007-2008 cycle (prior to then, was top-coded at 85).

Number of observed values: 56,017 (100%)

Mean (standard deviation): 47.41 (19.02)

Median (25th, 75th percentile): 31.00 (47.00, 63.00)

Number of missing values: 0 (0.00%)

demo_pregnant

Label: Pregnant

Description: Pregnancy status. This is defined by either self-report of being pregnant or a positive pregnancy test conducted during the study visit.

Number of observed values: 55,597 (99.3%)

Counts of observed values

No: 54,084 (97.3%)

Yes: 1,513 (2.72%)

Number of missing values: 420 (0.75%)

demo_gender

Label: Gender

Description: Self-reported gender

Number of observed values: 56,017 (100%)

Counts of observed values

Men: 27,242 (48.6%)

Women: 28,775 (51.4%)

Number of missing values: 0 (0.00%)

Blood pressure variables

Systolic and diastolic BP (SBP and DBP, respectively) were measured three times by trained and certified physicians. BP was measured using a mercury sphygmomanometer from 1999-2000 through 2015-2016 and using an oscillometric device in 2017-2020. The oscillometric SBP and DBP values were calibrated to the mercury device. Mean SBP and DBP levels were computed over all available

bp_sys_mean

Label: Systolic blood pressure (SBP), mm Hg

Description: Mean systolic blood pressure in mm Hg. This is based on the average of up to 3 readings. Participants were required to have at least one reading. Overall, >95% of participants with at least one systolic blood pressure reading had three readings. From 1999-2000 through 2015-2016, systolic blood pressure was measured using a mercury sphygmomanometer. In 2017-2020, systolic blood pressure was measured using an oscillometric device. The systolic blood pressure in 2017-2020 was calibrated to the mercury device by adding 1.5 mm Hg to the mean measured oscillometric value.

Number of observed values: 56,017 (100%)

Mean (standard deviation): 124.06 (19.34)

Median (25th, 75th percentile): 110.67 (120.67, 134.00)

Number of missing values: 0 (0.00%)

bp_dia_mean

Label: Diastolic blood pressure (DBP), mm Hg

Description: Mean diastolic blood pressure in mm Hg. This is based on the average of up to 3 readings. Participants were required to have at least one reading. Overall, >95% of participants with at least one diastolic blood pressure reading had three readings. From 1999-2000 through 2015-2016, diastolic blood pressure was measured using a mercury sphygmomanometer. In 2017-2020, diastolic blood pressure was measured using an oscillometric device. The diastolic blood pressure in 2017-2020 was calibrated to the mercury device by subtracting 1.3 mm Hg to the mean measured oscillometric value.

Number of observed values: 56,017 (100%)

Mean (standard deviation): 70.34 (12.10)

Median (25th, 75th percentile): 62.67 (70.37, 78.00)

Number of missing values: 0 (0.00%)

bp_cat_meds_excluded

Label: Blood pressure category

Description: Systolic/diastolic blood pressure <120/80, 120-129/<80, 130-139/80-89, 140-159/90-99, >= 160/100. All participants were placed in the category associated with a higher blood pressure (e.g., someone with systolic blood pressure of 150 mm Hg and diastolic blood pressure of 76 mm Hg was placed in the 140-159/90-99 mm Hg category)

Number of observed values: 56,017 (100%)

Counts of observed values

SBP <120 and DBP <80 mm Hg: 25,278 (45.1%)

SBP of 120 to <130 and DBP <80 mm Hg: 8,962 (16.0%)

SBP of 130 to <140 or DBP 80 to <90 mm Hg: 10,854 (19.4%)

SBP of 140 to <160 or DBP 90 to <100 mm Hg: 7,617 (13.6%)

SBP 160+ or DBP 100+ mm Hg: 3,306 (5.90%)

Number of missing values: 0 (0.00%)

bp_cat_meds_included

Label: Blood pressure category including antihypertensive medication use as a group

Description: Systolic/diastolic blood pressure <120/80, 120-129/<80, 130-139/80-89, 140-159/90-99, >= 160/100. Participants taking antihypertensive medication were placed in a separate category. Participants were placed in the category associated with higher blood pressure (e.g., someone with systolic blood pressure of 150 mm Hg and diastolic blood pressure of 76 mm Hg was placed in the 140-159/90-99 mm Hg category)

Number of observed values: 56,017 (100%)

Counts of observed values

SBP <120 and DBP <80 mm Hg: 22,359 (39.9%)

SBP of 120 to <130 and DBP < 80 mm Hg: 6,702 (12.0%)

SBP of 130 to <140 or DBP 80 to <90 mm Hg: 7,579 (13.5%)

SBP of 140 to <160 or DBP 90 to <100 mm Hg: 4,122 (7.36%)

SBP 160+ or DBP 100+ mm Hg: 1,508 (2.69%)

taking antihypertensive medications: 13,747 (24.5%)

Number of missing values: 0 (0.00%)

bp_control_jnc7

Label: Blood pressure control defined by the JNC7 guideline

Description: Systolic and diastolic blood pressure controlled to the levels recommended in the JNC7 guideline, systolic blood pressure < 140 mm Hg and diastolic blood pressure < 90 mm Hg except for those with diabetes or chronic kidney disease, where blood pressure control was defined by systolic blood pressure < 130 mm Hg and diastolic blood pressure < 80 mm Hg.

Number of observed values: 56,017 (100%)

Counts of observed values

No: 13,867 (24.8%)

Yes: 42,150 (75.2%)

Number of missing values: 0 (0.00%)

bp_control_accaha

Label: Blood pressure control defined by the 2017 ACC/AHA BP guideline

Description: Systolic and diastolic blood pressure controlled to the levels recommended in the 2017 ACC/AHA BP guideline, systolic blood pressure < 130 mm Hg and diastolic blood pressure < 80 mm Hg except for those >= 65 years of age without diabetes, chronic kidney disease, history of cardiovascular disease or 10-year predicted ASCVD risk >= 10% estimated using the Pooled Cohort risk equations. For this group, blood pressure control was defined as systolic blood pressure < 130 mm Hg

Number of observed values: 56,017 (100%)

Counts of observed values

No: 21,740 (38.8%)

Yes: 34,277 (61.2%)

Number of missing values: 0 (0.00%)

bp_control_140_90

Label: Blood pressure control (SBP < 140 mm Hg and DBP < 90 mm Hg)

Description: Systolic blood pressure < 140 mm Hg and diastolic blood pressure < 90 mm Hg

Number of observed values: 56,017 (100%)

Counts of observed values

No: 10,923 (19.5%)

Yes: 45,094 (80.5%)

Number of missing values: 0 (0.00%)

bp_control_130_80

Label: Blood pressure control (SBP < 130 mm Hg and DBP < 80 mm Hg)

Description: Systolic blood pressure < 130 mm Hg and diastolic blood pressure < 80 mm Hg

Number of observed values: 56,017 (100%)

Counts of observed values

No: 21,777 (38.9%)

Yes: 34,240 (61.1%)

Number of missing values: 0 (0.00%)

bp_uncontrolled_jnc7

Label: Uncontrolled blood pressure defined by the JNC7 guideline

Description: Systolic blood pressure >= 140 mm Hg or diastolic blood pressure >= 90 mm Hg for those without diabetes and without chronic kidney disease; Systolic blood pressure >= 130 mm Hg or diastolic blood pressure >= 80 mm Hg for those with diabetes or chronic kidney disease

Number of observed values: 56,017 (100%)

Counts of observed values

No: 42,150 (75.2%)

Yes: 13,867 (24.8%)

Number of missing values: 0 (0.00%)

bp_uncontrolled_accaha

Label: Uncontrolled blood pressure defined by the 2017 ACC/AHA BP guideline

Description: Systolic blood pressure >= 130 mm Hg or diastolic blood pressure >= 80 mm Hg except for adults >= 65 years of age without diabetes, chronic kidney disease, history of cardiovascular disease or 10-year predicted ASCVD risk >= 10%. For this group, uncontrolled blood pressure was defined as systolic blood pressure >= 130 mm Hg.

Number of observed values: 56,017 (100%)

Counts of observed values

No: 34,277 (61.2%)

Yes: 21,740 (38.8%)

Number of missing values: 0 (0.00%)

bp_uncontrolled_140_90

Label: Uncontrolled blood pressure (SBP >= 140 mm Hg or DBP >= 90 mm Hg)

Description: Systolic blood pressure >= 140 mm Hg or diastolic blood pressure >= 90 mm Hg

Number of observed values: 56,017 (100%)

Counts of observed values

No: 45,094 (80.5%)

Yes: 10,923 (19.5%)

Number of missing values: 0 (0.00%)

bp_uncontrolled_130_80

Label: Uncontrolled blood pressure (SBP >= 130 mm Hg or DBP >= 80 mm Hg)

Description: Systolic blood pressure >= 130 mm Hg or diastolic blood pressure >= 80 mm Hg

Number of observed values: 56,017 (100%)

Counts of observed values

No: 34,240 (61.1%)

Yes: 21,777 (38.9%)

Number of missing values: 0 (0.00%)

Hypertension variables

These variables are derived according to the 2017 American College of Cardiology/American Heart Association (ACC/AHA) BP guideline and the Seventh Joint National Committee (JNC7) BP guideline.

htn_jnc7

Label: Hypertension defined by the JNC7 guideline

Description: Hypertension defined by the JNC7 guideline, systolic blood pressure >= 140 mm Hg, diastolic blood pressure >= 90 mm Hg or self-reported antihypertensive medication use.

Number of observed values: 56,017 (100%)

Counts of observed values

No: 36,640 (65.4%)

Yes: 19,377 (34.6%)

Number of missing values: 0 (0.00%)

htn_accaha

Label: Hypertension defined by the 2017 ACC/AHA BP guideline

Description: Hypertension defined by the 2017 ACC/AHA blood pressure guideline, systolic blood pressure >= 130 mm Hg, diastolic blood pressure >= 80 mm Hg or self-reported antihypertensive medication use.

Number of observed values: 56,017 (100%)

Counts of observed values

No: 29,061 (51.9%)

Yes: 26,956 (48.1%)

Number of missing values: 0 (0.00%)

htn_aware

Label: Awareness of hypertension

Description: Self-report of a prior diagnosis of antihypertensive medication.

Number of observed values: 56,017 (100%)

Counts of observed values

No: 37,597 (67.1%)

Yes: 18,420 (32.9%)

Number of missing values: 0 (0.00%)

htn_resistant_jnc7

Label: Apparent resistant hypertension defined by the JNC7 guideline

Description: Taking 4 or more classes of antihypertensive medication, systolic blood pressure >= 140 mm Hg or diastolic blood pressure >= 90 mm Hg with the use of 3 classes of antihypertensive medication, or systolic blood pressure >= 130 mm Hg or diastolic blood pressure >= 80 mm Hg with the use of 3 classes of antihypertensive medication for those with diabetes or chronic kidney disease.

Number of observed values: 56,017 (100%)

Counts of observed values

No: 53,626 (95.7%)

Yes: 2,391 (4.27%)

Number of missing values: 0 (0.00%)

htn_resistant_accaha

Label: Apparent resistant hypertension defined by the 2017 ACC/AHA BP guideline

Description: Taking 4 or more classes of antihypertensive medication; systolic blood pressure >= 130 mm Hg or diastolic blood pressure >= 80 mm Hg with the use of 3 classes of antihypertensive medication for those < 65 years of age and those >= 65 years of age with diabetes, chronic kidney disease, history of cardiovascular disease or 10-year predicted ASCVD risk >= 10%; Systolic blood pressure >= 130 mm Hg with the use of 3 classes of antihypertensive medication for those >= 65 years of age without diabetes, chronic kidney disease, history of cardiovascular disease or 10-year predicted ASCVD risk >= 10%.

Number of observed values: 56,017 (100%)

Counts of observed values

No: 53,390 (95.3%)

Yes: 2,627 (4.69%)

Number of missing values: 0 (0.00%)

htn_resistant_accaha_thz

Label: Apparent resistant hypertension defined by the JNC7 guideline, requires thiazide diuretic

Description: Taking 4 or more classes of antihypertensive medication; systolic blood pressure >= 130 mm Hg or diastolic blood pressure >= 80 mm Hg with the use of 3 classes of antihypertensive medication for those < 65 years of age and those >= 65 years of age with diabetes, chronic kidney disease, history of cardiovascular disease or 10-year predicted ASCVD risk >= 10%; Systolic blood pressure >= 130 mm Hg with the use of 3 classes of antihypertensive medication for those >= 65 years of age without diabetes, chronic kidney disease, history of cardiovascular disease or 10-year predicted ASCVD risk >= 10%. To meet this definition of resistant hypertension, the participant had to be taking a thiazide diuretic.

Number of observed values: 56,017 (100%)

Counts of observed values

No: 54,484 (97.3%)

Yes: 1,533 (2.74%)

Number of missing values: 0 (0.00%)

htn_resistant_jnc7_thz

Label: Apparent resistant hypertension defined by the 2017 ACC/AHA BP guideline, requires thiazide diuretic

Description: Taking 4 or more classes of antihypertensive medication, systolic blood pressure >= 140 mm Hg or diastolic blood pressure >= 90 mm Hg with the use of 3 classes of antihypertensive medication, or systolic blood pressure >= 130 mm Hg or diastolic blood pressure >= 80 mm Hg with the use of 3 classes of antihypertensive medication for those with diabetes or chronic kidney disease. To meet this definition of resistant hypertension, the participant had to be taking a thiazide diuretic.

Number of observed values: 56,017 (100%)

Counts of observed values

No: 54,647 (97.6%)

Yes: 1,370 (2.45%)

Number of missing values: 0 (0.00%)

Antihypertensive medication variables

During the interview, the labels of medications that participants reported taking in the preceding 30 days were recorded.

bp_med_use

Label: Self-reported antihypertensive medication use

Description: Self-reported use of antihypertensive medication

Number of observed values: 56,017 (100%)

Counts of observed values

No: 42,270 (75.5%)

Yes: 13,747 (24.5%)

Number of missing values: 0 (0.00%)

bp_med_recommended_jnc7

Label: Antihypertensive medications recommended defined by the JNC7 guideline

Description: Systolic blood pressure >= 140 mm Hg or diastolic blood pressure >= 90 mm Hg; Systolic blood pressure >= 130 mm Hg or diastolic blood pressure >= 80 mm Hg for those with chronic kidney disease or diabetes. Those taking antihypertensive medications were considered to be recommended treatment.

Number of observed values: 56,017 (100%)

Counts of observed values

No: 35,266 (63.0%)

Yes: 20,751 (37.0%)

Number of missing values: 0 (0.00%)

bp_med_recommended_accaha

Label: Antihypertensive medications recommended by the 2017 ACC/AHA BP guideline

Description: Systolic blood pressure >= 130 mm Hg or diastolic blood pressure >= 80 mm Hg; Systolic blood pressure >= 130 mm Hg for those age >= 65 years without diabetes, chronic kidney disease, history of cardiovascular disease, or 10-year predicted atherosclerotic cardiovascular disease risk >= 10% by the pooled cohort risk equations. Those taking antihypertensive medications were considered to be recommended treatment.

Number of observed values: 56,017 (100%)

Counts of observed values

No: 34,003 (60.7%)

Yes: 22,014 (39.3%)

Number of missing values: 0 (0.00%)

bp_med_n_class

Label: Number of antihypertensive medication classes

Description: Number of antihypertensive medication classes being taken based on the pill bottle review.

Number of observed values: 55,600 (99.3%)

Counts of observed values

None: 39,489 (71.0%)

One: 6,885 (12.4%)

Two: 5,349 (9.62%)

Three: 2,710 (4.87%)

Four or more: 1,167 (2.10%)

Number of missing values: 417 (0.74%)

bp_med_combination

Label: Combination therapy

Description: Taking a single pill with two classes of antihypertensive medication.

Number of observed values: 55,600 (99.3%)

Counts of observed values

No: 52,523 (94.5%)

Yes: 3,077 (5.53%)

Number of missing values: 417 (0.74%)

bp_med_ace

Label: ACE inhibitors

Description: Taking an angiotensin converting enzyme inhibitor, defined using the pill bottle review. Drugs in this class included benazepril, captopril, enalapril, fosinopril, lisinopril, moexipril, perindopril, quinapril, ramipril, trandolapril

Number of observed values: 55,600 (99.3%)

Counts of observed values

No: 48,967 (88.1%)

Yes: 6,633 (11.9%)

Number of missing values: 417 (0.74%)

bp_med_aldo

Label: Aldosterone antagonists

Description: Taking an aldosterone antagonist, defined using the pill bottle review. Drugs in this class included eplerenone, spironolactone.

Number of observed values: 55,600 (99.3%)

Counts of observed values

No: 55,186 (99.3%)

Yes: 414 (0.74%)

Number of missing values: 417 (0.74%)

bp_med_alpha

Label: Alpha-1 blockers

Description: Taking an alpha blocker, defined using the pill bottle review. Drugs in this class included doxazosin, prazosin, terazosin.

Number of observed values: 55,600 (99.3%)

Counts of observed values

No: 54,831 (98.6%)

Yes: 769 (1.38%)

Number of missing values: 417 (0.74%)

bp_med_angioten

Label: Angiotensin receptor blockers

Description: Taking an angiotensin receptor blocker, defined using the pill bottle review. Drugs in this class included azilsartan, candesartan, eprosartan, irbesartan, losartan, olmesartan, telmisartan, valsartan.

Number of observed values: 55,600 (99.3%)

Counts of observed values

No: 52,083 (93.7%)

Yes: 3,517 (6.33%)

Number of missing values: 417 (0.74%)

bp_med_beta

Label: Beta blockers

Description: Taking a beta blocker. Drugs in this class included acebutolol, atenolol, betaxolol, bisoprolol, carvedilol, labetalol, metoprolol, nadolol, nebivolol, pindolol, penbutolol, propranolol, timolol.

Number of observed values: 55,600 (99.3%)

Counts of observed values

No: 49,527 (89.1%)

Yes: 6,073 (10.9%)

Number of missing values: 417 (0.74%)

bp_med_ccb

Label: Calcium channel blockers

Description: Taking a calcium channel blocker, defined using the pill bottle review. Drugs in this class included amlodipine, diltiazem, felodipine, isradipine, nicardipine, nifedipine, nisoldipine, verapamil.

Number of observed values: 55,600 (99.3%)

Counts of observed values

No: 50,738 (91.3%)

Yes: 4,862 (8.74%)

Number of missing values: 417 (0.74%)

bp_med_central

Label: Central alpha1 agonist and other centrally acting agents

Description: Taking a centrally acting agents, defined using the pill bottle review. Drugs in this class included clonidine, guanabenz, guanadrel, guanethidine, guanfacine, methyldopa, reserpine.

Number of observed values: 55,600 (99.3%)

Counts of observed values

No: 55,135 (99.2%)

Yes: 465 (0.84%)

Number of missing values: 417 (0.74%)

bp_med_renin_inhibitors

Label: Direct renin inhibitors

Description: Taking a renin inhibitor, defined using the pill bottle review. Drugs in this class included aliskiren.

Number of observed values: 55,600 (99.3%)

Counts of observed values

No: 55,577 (100%)

Yes: 23 (0.04%)

Number of missing values: 417 (0.74%)

bp_med_vasod

Label: Direct vasodilators

Description: Taking a vasodilator, defined using the pill bottle review. Drugs in this class included hydralazine, minoxidil.

Number of observed values: 55,600 (99.3%)

Counts of observed values

No: 55,331 (99.5%)

Yes: 269 (0.48%)

Number of missing values: 417 (0.74%)

bp_med_diur_loop

Label: Loop diuretics

Description: Taking a loop diuretic, defined using the pill bottle review. Drugs in this class included bumetanide, furosemide, torsemide, ethacrynic acid.

Number of observed values: 55,600 (99.3%)

Counts of observed values

No: 53,757 (96.7%)

Yes: 1,843 (3.31%)

Number of missing values: 417 (0.74%)

bp_med_diur_Ksparing

Label: Potassium sparing diuretics

Description: Taking a potassium-sparing diuretic, defined using the pill bottle review. Drugs in this class included amiloride, triamterene.

Number of observed values: 55,600 (99.3%)

Counts of observed values

No: 54,804 (98.6%)

Yes: 796 (1.43%)

Number of missing values: 417 (0.74%)

bp_med_diur_thz

Label: Thiazide or thiazide-type diuretics

Description: Taking a thiazide diuretic, defined using the pill bottle review. Drugs in this class included bendroflumethiazide, chlorthalidone, chlorothiazide, hydrochlorothiazide, hydroflumethiazide, indapamide, metolazone, methyclothiazide, trichlormethiazide.

Number of observed values: 55,600 (99.3%)

Counts of observed values

No: 50,592 (91.0%)

Yes: 5,008 (9.01%)

Number of missing values: 417 (0.74%)

Comorbidities variables

These variables describe comorbid conditions.

cc_smoke

Label: Smoking status

Description: Self-reported current cigarette smoking

Number of observed values: 53,077 (94.8%)

Counts of observed values

Never: 29,261 (55.1%)

Former: 12,912 (24.3%)

Current: 10,904 (20.5%)

Number of missing values: 2,940 (5.25%)

cc_bmi

Label: Body mass index, kg/m2

Description: Body mass index in kg/m2, estimated using the height and weight measured during the study examination.

Number of observed values: 54,862 (97.9%)

Counts of observed values

<25: 17,284 (31.5%)

25 to <30: 18,005 (32.8%)

30 to <35: 10,916 (19.9%)

35+: 8,657 (15.8%)

Number of missing values: 1,155 (2.06%)

cc_diabetes

Label: Prevalent diabetes

Description: HbA1c >= 6.5% or self-report of a prior diagnosis of diabetes with self-reported use of insulin or oral glucose-lowering medications.

Number of observed values: 56,017 (100%)

Counts of observed values

No: 48,778 (87.1%)

Yes: 7,239 (12.9%)

Number of missing values: 0 (0.00%)

cc_ckd

Label: Prevalent chronic kidney disease

Description: Estimated glomerular filtration rate < 60 ml/min/1.73 m2 or albumin-to-creatinine > 30 mg/g. Estimated glomerular filtration rate was calculated using the 2021 serum creatinine-based equation.

Number of observed values: 56,017 (100%)

Counts of observed values

No: 46,806 (83.6%)

Yes: 9,211 (16.4%)

Number of missing values: 0 (0.00%)

cc_cvd_mi

Label: History of myocardial infarction

Description: Self-reported history of myocardial infarction

Number of observed values: 52,087 (93.0%)

Counts of observed values

No: 49,829 (95.7%)

Yes: 2,258 (4.34%)

Number of missing values: 3,930 (7.02%)

cc_cvd_chd

Label: History of coronary heart disease

Description: Self-reported history of myocardial infarction or coronary heart disease

Number of observed values: 52,084 (93.0%)

Counts of observed values

No: 48,771 (93.6%)

Yes: 3,313 (6.36%)

Number of missing values: 3,933 (7.02%)

cc_cvd_stroke

Label: History of stroke

Description: Self-reported history of stroke

Number of observed values: 52,099 (93.0%)

Counts of observed values

No: 50,081 (96.1%)

Yes: 2,018 (3.87%)

Number of missing values: 3,918 (6.99%)

cc_cvd_ascvd

Label: History of ASCVD

Description: Self-reported history of coronary heart disease, myocardial infarction or stroke

Number of observed values: 56,017 (100%)

Counts of observed values

No: 51,268 (91.5%)

Yes: 4,749 (8.48%)

Number of missing values: 0 (0.00%)

cc_cvd_hf

Label: History of heart failure

Description: Self-reported history of heart failure

Number of observed values: 52,013 (92.9%)

Counts of observed values

No: 50,301 (96.7%)

Yes: 1,712 (3.29%)

Number of missing values: 4,004 (7.15%)

cc_cvd_any

Label: History of CVD

Description: Self-reported history of coronary heart disease, myocardial infarction, stroke or heart failure

Number of observed values: 56,017 (100%)

Counts of observed values

No: 50,646 (90.4%)

Yes: 5,371 (9.59%)

Number of missing values: 0 (0.00%)